+++ JavaScript/9-logger/api.di.container.js
@@ -1,5 +1,5 @@
 module.exports = {
-    common: Object.freeze({hash: require('./hash')}),
-    db: Object.freeze(require('./db')),
-    console: Object.freeze(require('./logger.di.container')),
+  common: Object.freeze({ hash: require('./hash') }),
+  db: Object.freeze(require('./db')),
+  console: Object.freeze(require('./logger.di.container')),
 };
