+++ JavaScript/9-logger/api/city.js
@@ -1,4 +1,4 @@
-'use strict';
-const {db} = require('../api.di.container');
+
+const { db } = require('../api.di.container');
 
 module.exports = db('city');
