+++ JavaScript/9-logger/body.js
@@ -1,10 +1,10 @@
-'use strict';
+
 
 const receiveArgs = async (req) => {
-    const buffers = [];
-    for await (const chunk of req) buffers.push(chunk);
-    const data = Buffer.concat(buffers).toString();
-    return JSON.parse(data);
+  const buffers = [];
+  for await (const chunk of req) buffers.push(chunk);
+  const data = Buffer.concat(buffers).toString();
+  return JSON.parse(data);
 };
 
 module.exports = receiveArgs;
