+++ JavaScript/9-logger/load.js
@@ -1,4 +1,4 @@
-'use strict';
+
 
 const fs = require('node:fs').promises;
 const vm = require('node:vm');
@@ -6,10 +6,10 @@ const vm = require('node:vm');
 const RUN_OPTIONS = { timeout: 5000, displayErrors: false };
 
 module.exports = async (filePath, sandbox) => {
-    const src = await fs.readFile(filePath, 'utf8');
-    const code = `'use strict';\n${src}`;
-    const script = new vm.Script(code);
-    const context = vm.createContext(Object.freeze({ ...sandbox }));
-    const exported = script.runInContext(context, RUN_OPTIONS);
-    return exported;
+  const src = await fs.readFile(filePath, 'utf8');
+  const code = `'use strict';\n${src}`;
+  const script = new vm.Script(code);
+  const context = vm.createContext(Object.freeze({ ...sandbox }));
+  const exported = script.runInContext(context, RUN_OPTIONS);
+  return exported;
 };
