+++ JavaScript/9-logger/logger.di.container.js
@@ -1,1 +1,3 @@
-module.exports = Object.freeze(require('./logger'));
+const config = require('./config');
+
+module.exports = Object.freeze(require(`./${config.logger}-logger`));
