+++ JavaScript/9-logger/main.js
@@ -1,4 +1,4 @@
-'use strict';
+
 
 const fsp = require('node:fs').promises;
 const path = require('node:path');
@@ -6,28 +6,19 @@ const config = require('./config.js');
 const server = require(`./${config.apiServer.transport}.js`);
 const staticServer = require('./static.js');
 const load = require('./load.js');
-const db = require('./db.js');
-const hash = require('./hash.js');
-const logger = require('./logger.js');
 
-
-const sandbox = {
-    console: Object.freeze(logger),
-    db: Object.freeze(db),
-    common: {hash},
-};
 const apiPath = path.join(process.cwd(), './api');
 const routing = {};
 
 (async () => {
-    const files = await fsp.readdir(apiPath);
-    for (const fileName of files) {
-        if (!fileName.endsWith('.js')) continue;
-        const filePath = path.join(apiPath, fileName);
-        const serviceName = path.basename(fileName, '.js');
-        routing[serviceName] = await load(filePath, sandbox);
-    }
+  const files = await fsp.readdir(apiPath);
+  for (const fileName of files) {
+    if (!fileName.endsWith('.js')) continue;
+    const filePath = path.join(apiPath, fileName);
+    const serviceName = path.basename(fileName, '.js');
+    routing[serviceName] = require(filePath);
+  }
 
-    staticServer('./static', config.staticServer.port);
-    server(routing, config.apiServer.port);
+  staticServer('./static', config.staticServer.port);
+  server(routing, config.apiServer.port);
 })();
