+++ JavaScript/9-logger/package.json
@@ -1,19 +1,19 @@
 {
-  "name": "layers-ioc",
-  "version": "1.0.0",
-  "description": "",
+  "name": "nodejs-2022-2023",
+  "version": "9.0.0",
+  "description": "How to use DDD, SOLID, GRASP and GoF in JavaScript and Node.js",
+  "license": "MIT",
+  "private": true,
   "main": "main.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/node": "^20.4.0"
+  "engines": {
+    "node": "14 || 16 || 18"
   },
   "dependencies": {
-    "pg": "^8.11.1",
-    "pino": "^8.17.1",
-    "ws": "^8.15.1"
+    "pg": "^8.8.0",
+    "pino": "^8.17.2",
+    "ws": "^8.12.0"
+  },
+  "devDependencies": {
+    "eslint": "^8.56.0"
   }
 }
