+++ JavaScript/9-logger/user.js
@@ -1,4 +1,4 @@
-'use strict';
+
 
 const db = require('./db.js');
 const hash = require('./hash.js');
@@ -6,26 +6,26 @@ const hash = require('./hash.js');
 const users = db('users');
 
 module.exports = {
-    async read(id) {
-        return await users.read(id, ['id', 'login']);
-    },
+  async read(id) {
+    return await users.read(id, ['id', 'login']);
+  },
 
-    async create({login, password}) {
-        const passwordHash = await hash(password);
-        return await users.create({login, password: passwordHash});
-    },
+  async create({ login, password }) {
+    const passwordHash = await hash(password);
+    return await users.create({ login, password: passwordHash });
+  },
 
-    async update(id, {login, password}) {
-        const passwordHash = await hash(password);
-        return await users.update(id, {login, password: passwordHash});
-    },
+  async update(id, { login, password }) {
+    const passwordHash = await hash(password);
+    return await users.update(id, { login, password: passwordHash });
+  },
 
-    async delete(id) {
-        return await users.delete(id);
-    },
+  async delete(id) {
+    return await users.delete(id);
+  },
 
-    async find(mask) {
-        const sql = 'SELECT login from users where login like $1';
-        return await users.query(sql, [mask]);
-    },
+  async find(mask) {
+    const sql = 'SELECT login from users where login like $1';
+    return await users.query(sql, [mask]);
+  },
 };
